Renamed cryptic 'massage' to less cryptic 'transform'. :-)
